Remove hyphens from UUID to satisfy mapping file ID requirements
Remove hyphens from UUID to satisfy mapping file ID requirements

I don't know if this is official or documented anywhere, but it looks
like the mapping file ID must be a UUID *without hyphens*. Uploading a
file with a UUID like 6ac7912c-d163-47a0-ae4e-f6d3ad3a21c7 results in
an API response like:

```
"error": {
    "code": 3,
    "message": "Invalid URL: Not a valid uuid."
}
```

Removing the hyphens makes it work.

I'm not sure how "generalizable" we want the code in this repository
to be. It seems to me that this target should specifically be used to
generate a UUID that can be used directly for a mapping file upload.
To that end, I think it is correct to follow the exact format that
makes this work, rather than make something purely generalizable.